+++ workspaces/bulk-import/plugins/bulk-import/src/components/AddRepositories/RepositoriesHeader.tsx
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-import type { ChangeEvent, MouseEvent } from 'react';
+import type { ChangeEvent } from 'react';
 
 import Checkbox from '@mui/material/Checkbox';
 import TableCell from '@mui/material/TableCell';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
 
 import { useGitlabConfigured } from '../../hooks';
+import { useImportFlow } from '../../hooks/useImportFlow';
 import { useTranslation } from '../../hooks/useTranslation';
-import { Order } from '../../types';
+import { ImportFlow } from '../../types';
 import { getRepositoriesListColumns } from '../Repositories/RepositoriesListColumns';
 import { getOrganizationsColumnHeader } from './OrganizationsColumnHeader';
 import { getRepositoriesColumnHeader } from './RepositoriesColumnHeader';
@@ -32,11 +32,8 @@ import { getReposSelectDrawerColumnHeader } from './ReposSelectDrawerColumnHeade
 
 export const RepositoriesHeader = ({
   onSelectAllClick,
-  order,
-  orderBy,
   numSelected,
   rowCount,
-  onRequestSort,
   isDataLoading,
   showOrganizations,
   showImportJobs,
@@ -44,9 +41,6 @@ export const RepositoriesHeader = ({
   isApprovalToolGitlab = false,
 }: {
   numSelected?: number;
-  onRequestSort: (event: MouseEvent<unknown>, property: any) => void;
-  order: Order;
-  orderBy: string | undefined;
   rowCount?: number;
   isDataLoading?: boolean;
   showOrganizations?: boolean;
@@ -56,23 +50,30 @@ export const RepositoriesHeader = ({
   onSelectAllClick?: (event: ChangeEvent<HTMLInputElement>) => void;
 }) => {
   const { t } = useTranslation();
-  const createSortHandler = (property: any) => (event: MouseEvent<unknown>) => {
-    onRequestSort(event, property);
-  };
 
   const gitlabConfigured = useGitlabConfigured();
+  const importFlow = useImportFlow();
+  const isScaffolderEnabled = importFlow === ImportFlow.Scaffolder;
 
   const getColumnHeader = () => {
     if (showOrganizations) {
       return getOrganizationsColumnHeader(isApprovalToolGitlab, t);
     }
     if (showImportJobs) {
-      return getRepositoriesListColumns(t, gitlabConfigured);
+      return getRepositoriesListColumns(
+        t,
+        gitlabConfigured,
+        isScaffolderEnabled,
+      );
     }
     if (isRepoSelectDrawer) {
       return getReposSelectDrawerColumnHeader(t);
     }
-    return getRepositoriesColumnHeader(isApprovalToolGitlab, t);
+    return getRepositoriesColumnHeader(
+      isApprovalToolGitlab,
+      t,
+      isScaffolderEnabled,
+    );
   };
 
   const tableCellStyle = () => {
@@ -99,7 +100,6 @@ export const RepositoriesHeader = ({
               padding: tableCellStyle(),
               fontWeight: '700',
             }}
-            sortDirection={orderBy === headCell.field ? order : 'asc'}
           >
             {index === 0 && !showOrganizations && !showImportJobs && (
               <Checkbox
@@ -123,14 +123,7 @@ export const RepositoriesHeader = ({
                 disabled={isDataLoading}
               />
             )}
-            <TableSortLabel
-              active={orderBy === headCell.field}
-              direction={orderBy === headCell.field ? order : 'asc'}
-              onClick={createSortHandler(headCell.field)}
-              disabled={headCell.sorting === false}
-            >
-              {headCell.title}
-            </TableSortLabel>
+            {headCell.title}
           </TableCell>
         ))}
       </TableRow>