+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization-common/src/permissions.ts
@@ -40,4 +40,20 @@ export const rosClusterProjectPermission = (
   });
 
 /** @public */
+export const costPluginReadPermission = createPermission({
+  name: 'cost.plugin',
+  attributes: { action: 'read' },
+});
+
+/** @public */
+export const costClusterSpecificPermission = (clusterName: string) =>
+  createPermission({
+    name: `cost.${clusterName}`,
+    attributes: { action: 'read' },
+  });
+
+/** @public */
+export const costPluginPermissions = [costPluginReadPermission];
+
+/** @public */
 export const rosPluginPermissions = [rosPluginReadPermission];